[libsae] Eliminate dependency on libconfig
meshd-nl80211 and the libsae themselves do use libconfig, but users of
libsae should not be required to link against libconfig.
