+++ linux/meshd-nl80211.c
@@ -60,6 +60,8 @@
 #include "service.h"
 #include "watch_ips.h"
 
+#include <libconfig.h>
+
 #define CIPHER_CCMP 0x000FAC04
 #define CIPHER_AES_CMAC 0x000FAC06
 
@@ -1521,6 +1523,43 @@ static int init(struct netlink_config_s *nlcfg, struct mesh_node *mesh)
 out:
     return 0;
 }
+
+static int
+sae_parse_libconfig (struct config_setting_t *sae_section, struct sae_config* config)
+{
+    struct config_setting_t *setting, *group;
+    char *pwd;
+
+    memset(config, 0, sizeof(struct sae_config));
+    config_setting_lookup_int(sae_section, "debug", (config_int_t *)&config->debug);
+    setting = config_setting_get_member(sae_section, "group");
+    if (setting != NULL) {
+        while (1) {
+            group = config_setting_get_elem(setting, config->num_groups);
+            if (!group)
+                break;
+            config->group[config->num_groups] =
+                config_setting_get_int_elem(setting, config->num_groups);
+            config->num_groups++;
+            if (config->num_groups == SAE_MAX_EC_GROUPS)
+                break;
+        }
+    }
+    if (config_setting_lookup_string(sae_section, "password", (const char **)&pwd)) {
+        strncpy(config->pwd, pwd, SAE_MAX_PASSWORD_LEN);
+        if (config->pwd[SAE_MAX_PASSWORD_LEN - 1] != 0) {
+            fprintf(stderr, "WARNING: Truncating password\n");
+            config->pwd[SAE_MAX_PASSWORD_LEN - 1] = 0;
+        }
+    }
+    config_setting_lookup_int(sae_section, "retrans", (config_int_t *)&config->retrans);
+    config_setting_lookup_int(sae_section, "lifetime", (config_int_t *)&config->pmk_expiry);
+    config_setting_lookup_int(sae_section, "thresh", (config_int_t *)&config->open_threshold);
+    config_setting_lookup_int(sae_section, "blacklist", (config_int_t *)&config->blacklist_timeout);
+    config_setting_lookup_int(sae_section, "giveup", (config_int_t *)&config->giveup_threshold);
+    return 0;
+}
+
 
 int main(int argc, char *argv[])
 {
